requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
google-genai
pdfplumber
scikit-learn
pandas
numpy
joblib